chore(curriculum): v0.157.1 DIP cleanup — relocate wide repository ports к application/usecases
Closes ADR-1 carry-forward from #269. Mechanical refactor — pure
DIP alignment per CLAUDE.md gate ("Repository interfaces — в
пакете-потребителе (usecase/), НЕ в domain/.").

Relocated 5 wide ports из internal/modules/curriculum/domain/repositories/
к internal/modules/curriculum/application/usecases/repository_interfaces.go:
CurriculumRepository, SectionRepository, DisciplineItemRepository,
BulkDisciplineItemsUnitOfWork, BulkDisciplineItemsTx.

Sentinels (ErrCurriculumNotFound etc.) and query DTOs (CurriculumListFilter,
CurriculumListResult, CurriculumYearSpecialtyAgg, DisciplineItemHoursAgg)
stay в domain/repositories — they're domain values consumed cross-module
by reports/annual per producer-module DTO ownership pattern.

13 qualified ref sites updated: 4 в infrastructure/persistence/bulk_unit_of_work_pg.go
(switch repositories.X → usecases.X + add import), 1 в bulk_edit_discipline_items_usecase.go
(narrow-port Begin return), 8 в bulk_edit_discipline_items_usecase_test.go
(collapse to unqualified, same-package).

Tier 2 absorbs (per feedback_tier2_absorb_same_release pattern):
- T2#1: plan doc count fix 9 → 13 sites (initial draft missed
  BulkDisciplineItemsTx as a separate identifier);
- T2#2: docstring slip в repository_interfaces.go header — now mentions
  Begin return type alongside accessor surface;
- T2#3: added 4 compile-time var _ usecases.<Port> = (*XxxPG)(nil)
  assertions в *_pg.go files so signature drift surfaces at infra
  compile site rather than only at DI wiring.

Verification:
- go build ./... green
- go test ./... full repo sweep green
- golangci-lint ./internal/modules/curriculum/... 0 issues
- AmE/BrE grep clean
- Reviewer pass SHIP (mean 8.78/10 pre-absorb)

Version bump: 0.157.0 → 0.157.1 (8 files via _tools/bump_version.sh).

Plan: docs/plans/2026-05-20-v0157-1-dip-cleanup.md
