+++ docs/plans/2026-05-20-v0157-1-dip-cleanup.md
@@ -1,0 +1,52 @@
+# v0.157.1 — Curriculum DIP Cleanup (Polish Patch)
+
+Closes ADR-1 carry-forward from #269 (v0.157.0 deferred к polish patch, mirror precedent v0.155 → v0.155.1).
+
+## Scope
+
+Mechanical refactor — relocate the curriculum module's **wide repository ports** from `internal/modules/curriculum/domain/repositories/` к `internal/modules/curriculum/application/usecases/` per CLAUDE.md DIP gate ("Repository interfaces — в пакете-потребителе (`usecase/`), НЕ в `domain/`. DIP по классике.").
+
+Zero behavior change. No new tests required (no behavioral surface added). Pure structural / dependency-arrow alignment.
+
+## Grep-verified facts (against `cbbcc428`)
+
+| Item | Count | Notes |
+|------|-------|-------|
+| Interfaces to relocate | **5** | `CurriculumRepository`, `SectionRepository`, `DisciplineItemRepository`, `BulkDisciplineItemsUnitOfWork`, `BulkDisciplineItemsTx` |
+| Qualified ref sites `repositories.<Interface>` | **13** | 4 в `infrastructure/persistence/bulk_unit_of_work_pg.go` (`Begin` return + `Items`/`Sections`/`Curricula` accessor returns), 1 в `application/usecases/bulk_edit_discipline_items_usecase.go` (narrow-port `Begin` return), 8 в `application/usecases/bulk_edit_discipline_items_usecase_test.go` (3 fake-impl comments + 1 fakeBulkTx comment + 3 fake Items/Sections/Curricula returns + 1 fake Begin return) |
+| Compile-time `var _` interface contracts | **0 pre-refactor / 4 post-refactor** | T2#3 absorbed: assertions added in `*_pg.go` files for `BulkDisciplineItemsUnitOfWorkPG` + the 3 wide repo PG types so signature drift surfaces at infra compile-time instead of DI wiring |
+| Sentinels staying в `domain/repositories/` | unchanged | `ErrCurriculumNotFound`, `ErrCurriculumCodeExists`, `ErrCurriculumVersionConflict`, `ErrSectionNotFound`, `ErrSectionVersionConflict`, `ErrBulkTxFinished` etc. |
+| Filter DTOs staying в `domain/repositories/` | unchanged | `CurriculumListFilter`, `CurriculumListResult`, `CurriculumYearSpecialtyAgg`, `SectionListFilter`, `DisciplineItemHoursAgg` etc. Cross-module consumer `reports/annual/application/usecases/annual_report_usecase.go` reads the `*Agg` DTOs — keeping them in producing module's domain mirrors `feedback_aggregate_dto_producer_module_ownership.md`. |
+
+NOTE: plan-doc factual drift in v0.157.0 originally claimed "52 import sites" — grep shows real scope is 13 qualified references, since most `domain/repositories` imports are for sentinels + DTOs (which stay). Counted ahead this time per `feedback_plan_doc_multi_release.md` — initial draft under-counted at 9 (missed `BulkDisciplineItemsTx` as a separate symbol from `BulkDisciplineItemsUnitOfWork`); refined to 13 after exhaustive grep ahead of refactor execution. Lesson reinforced: enumerate the EXACT identifier list before greppling.
+
+## ADR-1: Relocate curriculum wide ports к application/usecases/
+
+**Decision**: single aggregating file `internal/modules/curriculum/application/usecases/repository_interfaces.go` housing all 5 interface declarations. Mirroring к 4 files rejected — usecases/ dir would gain low-value mirror noise.
+
+**Mechanics**:
+1. Create new file with interface bodies copied verbatim (preserve docstrings).
+2. Strip interface blocks из `domain/repositories/{curriculum,section,discipline_item,bulk_unit_of_work}_repository.go` — sentinels + DTOs untouched.
+3. Update `infrastructure/persistence/bulk_unit_of_work_pg.go` — switch `repositories.<X>` → `usecases.<X>` (3 sites), add import.
+4. Update `application/usecases/bulk_edit_discipline_items_usecase_test.go` — drop qualifier (same-package, 6 sites collapse к unqualified).
+5. `go build ./...` + `go test ./internal/modules/curriculum/...` + `golangci-lint run ./internal/modules/curriculum/...`
+
+**Architectural arrow**: new infra → application/usecases arrow for the wide port. Pre-existing usecases → domain arrow unchanged (sentinels + DTOs still domain values). Classical DIP shape.
+
+## Out of scope (explicit deferrals)
+
+- Other modules (auth/users/announcements etc.) — addressed batches 2-4 if reviewers flag same drift.
+- Test refactor — fakes-renaming or narrow-port extraction (per-usecase ports already narrow; this patch touches only wide port location).
+
+## Definition of done
+
+- [ ] `go build ./...` green
+- [ ] `go test ./internal/modules/curriculum/...` green
+- [ ] `golangci-lint run ./internal/modules/curriculum/...` green
+- [ ] AmE/BrE grep clean
+- [ ] Reviewer pass (`superpowers:code-reviewer`) — single-pass SHIP expected (mechanical, ≥8/10 на каждой оси)
+- [ ] Version bumped 0.157.0 → 0.157.1 (8 files via `bash _tools/bump_version.sh 0.157.1`)
+- [ ] Release commit (`chore(release): v0.157.1 curriculum DIP cleanup`)
+- [ ] PR + admin-merge
+- [ ] Docs PR (roles-and-flows refresh) per sticky rule
+- [ ] Memory artifacts (`project_v0157_1_dip_cleanup.md` + MEMORY.md + chronicles vol 5 + handoff)
