+++ docs/swagger/swagger.yaml
@@ -1632,7 +1632,7 @@ info:
     name: MIT
     url: https://opensource.org/licenses/MIT
   title: Inf-Sys Secretary-Methodist API
-  version: 0.157.0
+  version: 0.157.1
 paths:
   /api/admin/notifications:
     post:
