+++ frontend/VERSION
@@ -1,1 +1,1 @@
-0.157.0
+0.157.1
