+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "inf-sys-secretary-methodist-frontend",
-  "version": "0.157.0",
+  "version": "0.157.1",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
