+++ internal/modules/curriculum/domain/repositories/section_repository.go
@@ -1,11 +1,6 @@
 package repositories
 
-import (
-	"context"
-	"errors"
-
-	"github.com/inf-sys-secretary-methodologist/inf-sys-secretary-methodist/internal/modules/curriculum/domain/entities"
-)
+import "errors"
 
 // ErrSectionNotFound signals that no Section row exists for the
 // requested id (or that the row was deleted between load and write).
@@ -25,36 +20,6 @@ var ErrSectionNotFound = errors.New("section: section not found")
 // "this section is gone").
 var ErrSectionVersionConflict = errors.New("section: version conflict")
 
-// SectionRepository is the persistence port for Section aggregates.
-// Implementations must satisfy the documented sentinel contract:
-// ErrSectionNotFound on missing rows; ErrSectionVersionConflict on
-// stale-version Update attempts.
-type SectionRepository interface {
-	// Save inserts a new Section and writes the generated id back onto
-	// the entity. version starts at 0 in the row; ID is set on success.
-	Save(ctx context.Context, s *entities.Section) error
-
-	// GetByID returns the Section with the given id or
-	// ErrSectionNotFound.
-	GetByID(ctx context.Context, id int64) (*entities.Section, error)
-
-	// ListByCurriculumID returns every Section attached to the given
-	// curriculum, ordered by (order_index ASC, created_at ASC, id ASC)
-	// for deterministic display. An empty result is not an error.
-	ListByCurriculumID(ctx context.Context, curriculumID int64) ([]*entities.Section, error)
-
-	// Update writes the (already-mutated) entity back. Implementations
-	// MUST enforce optimistic locking: WHERE id = ? AND version = ?.
-	// On RowsAffected == 0 the impl distinguishes via a follow-up
-	// existence check:
-	//   row missing entirely → ErrSectionNotFound
-	//   row exists but version stale → ErrSectionVersionConflict
-	// On success, the entity's version is bumped to reflect the new
-	// row state so callers see a consistent post-update view.
-	Update(ctx context.Context, s *entities.Section) error
-
-	// Delete removes the Section row by id. Returns ErrSectionNotFound
-	// if no row was deleted. CASCADE in migration 034 handles
-	// child-item cleanup automatically (DisciplineItem in v0.128.1+).
-	Delete(ctx context.Context, id int64) error
-}
+// The SectionRepository port itself lives в
+// internal/modules/curriculum/application/usecases/repository_interfaces.go
+// (DIP — interface lives with consumer). v0.157.1.
